matching service application-module test code complete

+++ matching-service/matching-application/src/main/java/shop/shportfolio/matching/application/ports/input/socket/OrderBookListener.java
@@ -3,5 +3,5 @@ package shop.shportfolio.matching.application.ports.input.socket;
 import shop.shportfolio.matching.application.dto.orderbook.OrderBookBithumbDto;
 
 public interface OrderBookListener {
-    void onOrderBookReceived(OrderBookBithumbDto dto, double marketItemTick);
+    void onOrderBookReceived(OrderBookBithumbDto dto, Long marketItemTick);
 }
